move echo mocks
