+++ Sources/WalletConnectEcho/EchoClient.swift
@@ -17,3 +17,18 @@ public class EchoClient: EchoClientProtocol {
     }
 #endif
 }
+
+
+#if DEBUG
+final class EchoClientMock: EchoClientProtocol {
+    var registedCalled = false
+
+    func register(deviceToken: Data) async throws {
+        registedCalled = true
+    }
+
+    func register(deviceToken: String) async throws {
+        registedCalled = true
+    }
+}
+#endif
