--- Tests/TestingUtils/Mocks/EchoClientMock.swift
@@ -1,12 +1,0 @@
-
-import Foundation
-
-class EchoClientMock: EchoClientProtocol {
-    func register(deviceToken: Data) async throws {
-
-    }
-
-    func register(deviceToken: String) async throws {
-
-    }
-}
